--- client/src/pages/GamePlaying/GamePlaying.tsx
@@ -1,9 +1,0 @@
-const page = () => {
-    return (
-        <div>
-            게임 플레잉 페이지
-        </div>
-    )
-}
-
-export default page
