--- client/src/pages/MakeRoom.tsx
@@ -1,9 +1,0 @@
-const MakeRoom = () => {
-    return (
-        <div>
-
-        </div>
-    )
-}
-
-export default MakeRoom
